+++ src/main/java/me/marin/lockout/mixin/server/StatusEffectMixin.java
@@ -38,6 +38,7 @@ public class StatusEffectMixin {
                     lockout.completeGoal(goal, player);
                 }
             }
+            
         }
     }